requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
da-vinci==0.3.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1
django-appconf==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2.1
django-filter==2.3.0
-e git://github.com/kmmbvnr/django-fsm.git@2550746dc038fe1fb8eeb77bf260076590154722#egg=django_fsm
django-heroku==0.3.1
django-imagekit==4.0.2
django-push-notifications==2.0.0
django-smtp-ssl==1.0
django-thumbnails==0.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.7.0
drf-yasg==1.20.0
fcm-django==0.3.10
future==0.18.2
gcloud==0.18.3
geocoder==1.38.1
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
html5lib==1.1
http-ece==1.1.0
httplib2==0.19.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.8
MarkupSafe==2.0.1
Naked==0.1.31
oauth2client==4.1.3
packaging==20.9
pilkit==2.0
Pillow==7.2.0
protobuf==3.15.0
psycopg2==2.8.5
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pyfcm==1.5.1
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
Pyrebase4==4.4.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywebpush==1.13.0
PyYAML==5.4.1
ratelim==0.1.6
redis==3.5.3
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
shellescape==3.8.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
